Move Monero related classes to xmr package
